change ApiErrorHandler and make some refactoring

+++ src/main/java/omsu/imit/schedule/service/ActivityTypeService.kt
@@ -1,11 +1,12 @@
 package omsu.imit.schedule.service
 
+import omsu.imit.schedule.exception.ErrorCode
+import omsu.imit.schedule.exception.NotFoundException
 import omsu.imit.schedule.model.ActivityType
 import omsu.imit.schedule.repository.ActivityTypeRepository
 import omsu.imit.schedule.requests.CreateActivityTypeRequest
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import javax.persistence.EntityNotFoundException
 
 @Service
 class ActivityTypeService
@@ -22,7 +23,9 @@ constructor(private val activityTypeRepository: ActivityTypeRepository) {
     fun getActivityTypeById(activityTypeId: Int): ActivityType? {
         return activityTypeRepository
                 .findById(activityTypeId)
-                .orElseThrow { EntityNotFoundException(String.format("Activity type  with id=%d not found", activityTypeId)) };
+                .orElseThrow {
+                    NotFoundException(ErrorCode.ACTIVITY_TYPE_NOT_EXISTS, activityTypeId.toString())
+                }
     }
 
     fun getAllActivityTypes(): Any {
@@ -30,6 +33,8 @@ constructor(private val activityTypeRepository: ActivityTypeRepository) {
     }
 
     fun deleteActivityType(activityTypeId: Int) {
+        if (!activityTypeRepository.existsById(activityTypeId))
+            throw NotFoundException(ErrorCode.ACTIVITY_TYPE_NOT_EXISTS, activityTypeId.toString())
         activityTypeRepository.deleteById(activityTypeId)
     }
