+++ src/main/java/omsu/imit/schedule/service/ChairService.kt
@@ -1,21 +1,25 @@
 package omsu.imit.schedule.service
 
+import omsu.imit.schedule.exception.ErrorCode
+import omsu.imit.schedule.exception.NotFoundException
+import omsu.imit.schedule.model.Chair
 import omsu.imit.schedule.repository.ChairRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import javax.persistence.EntityNotFoundException
 
 @Service
 class ChairService
 @Autowired
 constructor(private val chairRepository: ChairRepository) {
 
-    fun getChair(chairId: Int): Any {
+    fun getChair(chairId: Int): Chair {
         return chairRepository.findById(chairId)
-                .orElseThrow { EntityNotFoundException(String.format("Chair with id=%d not found", chairId)) }
+                .orElseThrow { NotFoundException(ErrorCode.CHAIR_NOT_EXISTS, chairId.toString()) }
     }
 
     fun deleteChair(chairId: Int) {
+        if (!chairRepository.existsById(chairId))
+            throw  NotFoundException(ErrorCode.CHAIR_NOT_EXISTS, chairId.toString())
         chairRepository.deleteById(chairId)
     }
