+++ src/main/java/omsu/imit/schedule/service/DisciplineService.kt
@@ -1,11 +1,12 @@
 package omsu.imit.schedule.service
 
+import omsu.imit.schedule.exception.ErrorCode
+import omsu.imit.schedule.exception.NotFoundException
 import omsu.imit.schedule.model.Discipline
 import omsu.imit.schedule.repository.DisciplineRepository
 import omsu.imit.schedule.requests.DisciplineRequest
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import javax.persistence.EntityNotFoundException
 
 @Service
 class DisciplineService
@@ -21,7 +22,7 @@ constructor(private val disciplineRepository: DisciplineRepository) {
 
     fun getDiscipline(disciplineId: Int): Discipline {
         return disciplineRepository.findById(disciplineId)
-                .orElseThrow { EntityNotFoundException(String.format("Discipline with id=%d not found", disciplineId)) }
+                .orElseThrow { NotFoundException(ErrorCode.DISCIPLINE_NOT_EXISTS, disciplineId.toString()) }
     }
 
     fun getAllDisciplines(): MutableList<Discipline> {
@@ -37,6 +38,8 @@ constructor(private val disciplineRepository: DisciplineRepository) {
     }
 
     fun deleteDiscipline(disciplineId: Int) {
+        if (!disciplineRepository.existsById(disciplineId))
+            throw NotFoundException(ErrorCode.DISCIPLINE_NOT_EXISTS, disciplineId.toString());
         disciplineRepository.deleteById(disciplineId)
     }
