+++ src/main/java/omsu/imit/schedule/service/FacultyService.kt
@@ -1,5 +1,7 @@
 package omsu.imit.schedule.service
 
+import omsu.imit.schedule.exception.ErrorCode
+import omsu.imit.schedule.exception.NotFoundException
 import omsu.imit.schedule.model.Building
 import omsu.imit.schedule.model.Faculty
 import omsu.imit.schedule.repository.BuildingRepository
@@ -7,7 +9,6 @@ import omsu.imit.schedule.repository.FacultyRepository
 import omsu.imit.schedule.requests.CreateFacultyRequest
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import javax.persistence.EntityNotFoundException
 
 @Service
 class FacultyService
@@ -16,7 +17,7 @@ class FacultyService
     fun createFaculty(request: CreateFacultyRequest): Faculty {
         val building: Building = buildingRepository
                 .findById(request.buildingId)
-                .orElseThrow { EntityNotFoundException(String.format("Building with id=%d not found", request.buildingId)) }
+                .orElseThrow { NotFoundException(ErrorCode.BUILDING_NOT_EXISTS, request.buildingId.toString()) }
 
         val faculty = Faculty(building, request.name)
         facultyRepository.save(faculty)
@@ -26,7 +27,7 @@ class FacultyService
 
     fun getFaculty(facultyId: Int): Faculty {
         return facultyRepository.findById(facultyId)
-                .orElseThrow { EntityNotFoundException(String.format("Faculty with id=%d not found", facultyId)) }
+                .orElseThrow { NotFoundException(ErrorCode.FACULTY_NOT_EXISTS, facultyId.toString()) }
     }
 
     fun getAllFaculties(): List<Faculty> {
@@ -34,6 +35,8 @@ class FacultyService
     }
 
     fun deleteFaculty(facultyId: Int) {
+        if (!facultyRepository.existsById(facultyId))
+            throw NotFoundException(ErrorCode.FACULTY_NOT_EXISTS, facultyId.toString())
         facultyRepository.deleteById(facultyId)
     }
 }
